refactor(vscode extension) split into separate modules
